requirements.txt
yfinance==0.1.55
azure-functions
pyodbc
requests
pandas
slackclient
requests_futures